requirements.txt
streamlit
google-generativeai
gtts
speechrecognition
pyaudio
audio_recorder_streamlit
